requirements.txt
asgiref==3.3.1
autopep8==1.5.5
cached-property==1.5.2
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
enum-compat==0.0.3
flake8==3.9.0
mccabe==0.6.1
pep8==1.7.1
Pillow==8.1.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-slugify==4.0.1
pytils==0.3
pytz==2021.1
reportlab==3.5.59
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
